+++ media-picker/src/wasmJsMain/kotlin/io/github/aashutosh/mediapicker/internal/ContextDiscovery.wasmJs.kt
@@ -1,0 +1,18 @@
+/*
+ * Copyright 2026 Aashutosh Kumar
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ */
+package io.github.aashutosh.mediapicker.internal
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+
+internal object WebPlatformContext : InternalPlatformContext
+
+@Composable
+internal actual fun discoverPlatformContext(): InternalPlatformContext? = remember { WebPlatformContext }