+++ samples/composeApp/src/androidMain/kotlin/io/github/aashutosh/mediapicker/sample/MainActivity.kt
@@ -12,13 +12,10 @@ package io.github.aashutosh.mediapicker.sample
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import io.github.aashutosh.mediapicker.AndroidPlatformContext
-import io.github.aashutosh.mediapicker.initializeMediaPicker
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        initializeMediaPicker(AndroidPlatformContext(this))
         setContent { App() }
     }
 }